Add "efficiency" and "utility"
Efficiency is the percentage of getworks compared to the amount of
work requested from the pool. It can go higher than 100% in case
more shares than getworks were found. Some pools prefer miners to
have a high efficiency; CPU miners likely exhibit a low efficiency.

Utility is the number of shares found per minute, since the miner
was started. It is another way to describe the effectiveness of
a miner.