driver-ztex: search the complete noncerange based on the actual speed
search the complete noncerange until the range to the end is more
work than at the last round. doing one more round would mean we
would have a overrun, which is a waste.

with actual ztex boards this means that a new getwork is needed
every 19 seconds in general and not every 10 seconds (without
rollntime).